Moving most of generate_dictionary into the library.
We will use this in followups to script auto-generation of dictionaries during benchmarking.
